requirements.txt
requests>=2.32.0
google-auth>=2.35.0
google-auth-oauthlib>=1.2.0